dest file path check

+++ src/com/kostyukov/Main.java
@@ -14,7 +14,7 @@ public class Main
 			Path sourceFile = FileSystems.getDefault().getPath("Examples/file1.txt");
 			Path destFIle = FileSystems.getDefault().getPath("Examples/file1_copy.txt");
 			
-			if (Files.exists(sourceFile))
+			if (Files.exists(sourceFile) && Files.notExists(destFIle))
 			{
 				Files.copy(sourceFile,destFIle);
 				System.out.println("File: " + sourceFile.toAbsolutePath() + " is copied successfully");
